Testing with other distros.

+++ package/make_src_deb.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 
+if [ -z "$1" ]
+  then
+    echo "Please specify an Ubuntu release (ie precise)"
+    exit 1
+fi
+
 SVNVER=`svn info ../ | grep Revision | awk '{ print $2 }'`
 echo $SVNVER
 
-BUILD_FOLDER=qstlink2-0.$SVNVER
+BUILD_FOLDER=qstlink2-0.$SVNVER~$1
 ORIG_FILE=$BUILD_FOLDER.orig.tar.gz
 
 rm -f $ORIG_FILE
-rm -rf $BUILD_FOLDER *.build *.changes *.ppa* *.dsc
+#rm -rf $BUILD_FOLDER *.build *.changes *.ppa* *.dsc
+rm -rf $BUILD_FOLDER
 mkdir $BUILD_FOLDER
 
 shopt -s extglob
@@ -24,7 +31,7 @@ echo "" | dh_make -n --single -e fabien.poussin@gmail.com -c gpl3
 cp -v ../debian/* debian/
 rm -vf debian/*.ex debian/*.EX debian/ex.*
 
-sed -i -e's/unstable/precise/' debian/changelog
+sed -i -e"s/unstable/$1/" debian/changelog
 
 debuild -j4 -S -sa
